Revert "fix(code-mode): carry owner effect receipts (#132146)"
This reverts commit a920626fc6d4f4001e0a5f723ced355467434257.
