+++ src/agents/code-mode.bridge.host-denial.test.ts
@@ -27,7 +27,6 @@ import { createExecTool } from "./bash-tools.exec-run.js";
 import type { ExecToolDefaults } from "./bash-tools.exec-types.js";
 import * as codeModeBridge from "./code-mode-bridge.js";
 import { consumeRepairableCodeModeFailure } from "./code-mode-repair-provenance.js";
-import { isCodeModeBridgeRepairEligible } from "./code-mode-state.js";
 import { createSubscribedCodeModeHarness } from "./code-mode.bridge.lifecycle.test-support.js";
 import { applyCodeModeCatalog } from "./code-mode.js";
 import {
@@ -37,7 +36,6 @@ import {
   testing,
   waitUntilCompleted,
 } from "./code-mode.test-support.js";
-import { consumeToolEffectReceipt } from "./tool-effect-receipt.js";
 import { consumeTrustedToolNoStartError } from "./tool-result-error.js";
 import { createToolTerminalObserver } from "./tool-terminal-outcome.js";
 import { jsonResult, ToolInputError } from "./tools/common.js";
@@ -547,13 +545,13 @@ describe("Code Mode subscribed host denial", () => {
           testing.activeRuns.get(String(details.runId)),
           "real parked history",
         );
-        expect(isCodeModeBridgeRepairEligible(parked.bridgeDispatch)).toBe(false);
+        expect(parked.bridgeDispatch.potentiallyMutatingDispatches).toBeGreaterThanOrEqual(1);
         details = resultDetails(
           await expectDefined(harness.tools[1], "wait").execute("resume-denial", {
             runId: details.runId,
           }),
         );
-        expect(isCodeModeBridgeRepairEligible(parked.bridgeDispatch)).toBe(false);
+        expect(parked.bridgeDispatch.potentiallyMutatingDispatches).toBeGreaterThanOrEqual(1);
       }
       if (order === "late-settlement") {
         expect(details.status).toBe("waiting");
@@ -561,7 +559,7 @@ describe("Code Mode subscribed host denial", () => {
           testing.activeRuns.get(String(details.runId)),
           "pending mutation",
         );
-        expect(isCodeModeBridgeRepairEligible(pending.bridgeDispatch)).toBe(false);
+        expect(pending.bridgeDispatch.potentiallyMutatingDispatches).toBe(1);
         expect(consumeRepairableCodeModeFailure(details)).toBe(false);
         await expect(fs.readFile(marker, "utf8")).rejects.toMatchObject({ code: "ENOENT" });
         releaseLateMutation.resolve();
@@ -572,7 +570,7 @@ describe("Code Mode subscribed host denial", () => {
           }),
         );
         await vi.waitFor(() => expect(harness.subscription.getItemLifecycle().activeCount).toBe(0));
-        expect(isCodeModeBridgeRepairEligible(pending.bridgeDispatch)).toBe(false);
+        expect(pending.bridgeDispatch.potentiallyMutatingDispatches).toBe(1);
       }
       details = await harness.complete(details);
       expect(details).toMatchObject({ status: "failed", bridgeDispatchStarted: true });
@@ -725,9 +723,9 @@ describe("Code Mode subscribed host denial", () => {
           const reused = await harness.runToCompletion();
           expect(consumeRepairableCodeModeFailure(reused)).toBe(false);
         } else {
-          expect(consumeToolEffectReceipt(first)).toEqual({ state: "not_started" });
+          expect(consumeTrustedToolNoStartError(first)).toBe(true);
         }
-        expect(consumeToolEffectReceipt(first)).toBeUndefined();
+        expect(consumeTrustedToolNoStartError(first)).toBe(false);
         expect(harness.spawn).not.toHaveBeenCalled();
       } finally {
         harness.dispose();
