+++ src/agents/embedded-agent-subscribe.handlers.tools.completion.ts
@@ -234,7 +234,6 @@ export async function handleToolExecutionEnd(
     arguments: startArgs,
     ...(meta ? { meta } : {}),
     executionStarted,
-    replaySafe: callSummary.replaySafe,
     outcome: isToolError ? "failure" : "success",
     ...(callSummary.ownerKey
       ? {
@@ -716,5 +715,5 @@ export async function handleToolExecutionEnd(
         ctx.log.warn(`after_tool_call hook failed: tool=${toolName} error=${String(err)}`);
       });
   }
-  return { executionStarted: terminal.executionStarted, effectReceipt: terminal.effectReceipt };
+  return { executionStarted: terminal.executionStarted };
 }
