+++ src/agents/embedded-agent-subscribe.ts
@@ -38,8 +38,10 @@ import { buildToolLifecycleErrorResult } from "./embedded-agent-tool-results.js"
 import { stripDowngradedToolCallText } from "./embedded-agent-utils.js";
 import type { AgentRunTimeoutPhase } from "./run-timeout-attribution.js";
 import type { AgentMessage } from "./runtime/index.js";
-import { registerToolEffectReceipt } from "./tool-effect-receipt.js";
-import { consumeTrustedToolNoStartError } from "./tool-result-error.js";
+import {
+  consumeTrustedToolNoStartError,
+  registerTrustedToolNoStartError,
+} from "./tool-result-error.js";
 import { hasNonzeroUsage, normalizeUsage, type UsageLike } from "./usage.js";
 
 const embeddedLog = createSubsystemLogger("agent/embedded");
@@ -627,7 +629,7 @@ export function subscribeEmbeddedAgentSession(params: SubscribeEmbeddedAgentSess
       };
       try {
         const result = await toolParams.execute(onImplementationStart);
-        const terminal = await handleToolExecutionEnd(ctx, {
+        await handleToolExecutionEnd(ctx, {
           type: "tool_execution_end",
           toolName: toolParams.toolName,
           toolCallId: toolParams.toolCallId,
@@ -636,10 +638,10 @@ export function subscribeEmbeddedAgentSession(params: SubscribeEmbeddedAgentSess
           result,
           hideFromChannelProgress: toolParams.hideFromChannelProgress,
         } as never);
-        return registerToolEffectReceipt(result, terminal.effectReceipt);
+        return result;
       } catch (error) {
         const trustedNoStart = consumeTrustedToolNoStartError(error);
-        const terminal = await handleToolExecutionEnd(ctx, {
+        await handleToolExecutionEnd(ctx, {
           type: "tool_execution_end",
           toolName: toolParams.toolName,
           toolCallId: toolParams.toolCallId,
@@ -648,10 +650,12 @@ export function subscribeEmbeddedAgentSession(params: SubscribeEmbeddedAgentSess
           result: buildToolLifecycleErrorResult(error),
           hideFromChannelProgress: toolParams.hideFromChannelProgress,
         } as never);
-        const receipt = trustedNoStart
-          ? ({ state: "not_started" } as const)
-          : terminal.effectReceipt;
-        throw registerToolEffectReceipt(error, receipt);
+        // Operation-owned no-start proof survives generic implementation entry.
+        // Only relay the same error after completion succeeds; replacements cannot inherit it.
+        if (trustedNoStart) {
+          registerTrustedToolNoStartError(error);
+        }
+        throw error;
       }
     },
     unsubscribe,
