+++ src/agents/tool-search-types.ts
@@ -62,8 +62,6 @@ export type ToolSearchCatalogToolExecutor = (params: {
   sourceName?: string;
   toolCallId: string;
   parentToolCallId?: string;
-  /** Exact registered-instance classification resolved by the catalog owner. */
-  replaySafe?: boolean;
   input: unknown;
   signal?: AbortSignal;
   onUpdate?: AgentToolUpdateCallback;
